started refactoring from first level of schema
