--- src/declare4py/constraint_checkers/__init__.py
@@ -1,4 +1,0 @@
-from .existence import *
-from .relation import *
-from .negative_relation import *
-from .choice import *
