--- src/declare4py/enums/__init__.py
@@ -1,1 +1,0 @@
-from .mp_constants import *
