--- src/declare4py/models/__init__.py
@@ -1,2 +1,0 @@
-from .checker_result import *
-from .decl_model import *
