+++ src/declare4py/monitoring.py
@@ -1,0 +1,1 @@
+class Monitoring: None
