+++ src/declare4py/old_structure/checker_result.py
