+++ src/declare4py/old_structure/choice.py
