+++ src/declare4py/old_structure/negative_relation.py
