+++ src/declare4py/old_structure/relation.py
