--- src/declare4py/parsers/__init__.py
@@ -1,1 +1,0 @@
-from .decl_parser import *
